moved modal state to different zustand file
