+++ src/components/PerformanceScoreEditor.tsx
@@ -1,4 +1,4 @@
-import { useAppState } from "../state/state";
+import { useAppState } from "../state/app.state";
 import { PerformanceScore } from "../business/common";
 
 export const PerformanceScoreEditor = () => {
