+++ src/components/RoleEditor.tsx
@@ -1,4 +1,4 @@
-import { useAppState } from "../state/state";
+import { useAppState } from "../state/app.state";
 
 export const RoleEditor = () => {
   const inputEnabled = useAppState((state) => state.inputEnabled);
